[refactor] #22 후보자 조회 api를 client패키지로 분리
